# docs/swagger.yaml
basePath: /
definitions:
  controllers.FindRequest:
    properties:
      attributes:
        items:
          type: string
        type: array
      searchBase:
        type: string
      searchFilter:
        type: string
    type: object
  controllers.FindResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/ldap_connector.LdapEntry'
        type: array
      message:
        type: string
      status:
        type: integer
    type: object
  controllers.LoginRequest:
    properties:
      bindDN:
        type: string
      bindPassword:
        type: string
      ldapURL:
        type: string
    type: object
  controllers.LoginResponse:
    properties:
      data:
        properties:
          session_token:
            type: string
        type: object
      message:
        type: string
      status:
        type: integer
    type: object
  controllers.UserAuthRequest:
    properties:
      password:
        type: string
      userDn:
        type: string
    required:
    - password
    - userDn
    type: object
  controllers.UserAuthResponse:
    properties:
      data:
        type: object
      message:
        type: string
      status:
        type: integer
    type: object
  ldap_connector.LdapEntry:
    properties:
      attributes:
        items:
          $ref: '#/definitions/ldap_connector.LdapEntryAttribute'
        type: array
      dn:
        type: string
    type: object
  ldap_connector.LdapEntryAttribute:
    properties:
      byteValues:
        description: ByteValues contain the raw values of the attribute
        items:
          items:
            type: integer
          type: array
        type: array
      name:
        description: Name is the name of the attribute
        type: string
      values:
        description: Values contain the string values of the attribute
        items:
          type: string
        type: array
    type: object
host: localhost:8082
info:
  contact:
    email: cihantaylan@cihantaylan.com
    name: API Support
    url: https://cihantaylan.com
  description: This is a go ldap rest API Documentation.
  title: Go Ldap Rest API
  version: "1.0"
paths:
  /api/v1/ldap/find:
    post:
      consumes:
      - application/json
      description: 'EG;<br>ByRG81IDDPQFY9+9dSaWFKIA3Xp1vZhrpCjCg4XXR7gnNxLM9SvgTK1PFKMrsdE5s4mNRSIo8qJhzeZAdMi5zQfAhJOV8FDdmEs=<br>SearchBase:
        dc=example,dc=com<br>SearchFilter: (objectClass=person)'
      operationId: Find
      parameters:
      - description: Your Auth Token
        in: header
        name: Token
        required: true
        type: string
      - description: Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.FindRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.FindResponse'
      summary: Find
      tags:
      - Ldap
  /api/v1/ldap/login:
    post:
      consumes:
      - application/json
      description: 'EG;<br>LdapURL: ldap://ldap.forumsys.com:389<br>BindDN: cn=read-only-admin,dc=example,dc=com<br>BindPassword:
        password'
      operationId: Login
      parameters:
      - description: Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.LoginResponse'
      summary: Login
      tags:
      - Ldap
  /api/v1/ldap/user-auth:
    post:
      consumes:
      - application/json
      description: 'EG;<br>ByRG81IDDPQFY9+9dSaWFKIA3Xp1vZhrpCjCg4XXR7gnNxLM9SvgTK1PFKMrsdE5s4mNRSIo8qJhzeZAdMi5zQfAhJOV8FDdmEs=<br>username:
        cn=read-only-admin,dc=example,dc=com<br>password: password'
      operationId: UserAuth
      parameters:
      - description: Your Auth Token
        in: header
        name: Token
        required: true
        type: string
      - description: Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.UserAuthRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.UserAuthResponse'
      summary: UserAuth
      tags:
      - Ldap
swagger: "2.0"